LFLT-534 Files cannot be uploaded to LSRS via Jersey REST client  - Refactored Formidable middleware and moved into separate module
