+++ test/lsrs/web/utility/middleware.test.ts
@@ -1,7 +1,6 @@
 import {isUUID} from "class-validator";
 import {Request, Response} from "express";
-import {IncomingForm} from "formidable";
-import sinon, {SinonStub, SinonStubbedInstance} from "sinon";
+import sinon, {SinonStubbedInstance} from "sinon";
 import {
     ConflictingResourceError,
     GenericError,
@@ -13,17 +12,12 @@ import {UploadedFile} from "../../../../src/lsrs/core/model/uploaded-file";
 import LoggerFactory from "../../../../src/lsrs/helper/logger-factory";
 import {InvalidRequestError} from "../../../../src/lsrs/web/error/api-error-types";
 import {ConstraintViolation, HttpStatus} from "../../../../src/lsrs/web/model/common";
-import {
-    errorHandlerMiddleware,
-    formidableUploadMiddleware,
-    requestTrackingMiddleware
-} from "../../../../src/lsrs/web/utility/middleware";
+import {errorHandlerMiddleware, requestTrackingMiddleware} from "../../../../src/lsrs/web/utility/middleware";
 
 describe("Unit tests for Express middleware functions", () => {
 
     let requestStub: Request;
     let responseStub: SinonStubbedInstance<Response>;
-    let parseStub: SinonStub;
 
     beforeEach(() => {
         requestStub = {body: {}} as unknown as Request;
@@ -63,72 +57,6 @@ describe("Unit tests for Express middleware functions", () => {
                     ? {message: scenario.error.message, violations: violations}
                     : {message: scenario.error.message});
             });
-        });
-    });
-
-    describe("Test scenarios for #formidableUploadMiddleware", () => {
-
-        afterEach(() => {
-            parseStub.restore();
-        });
-
-        it("should parse file upload form and pass fields to Express Request object", async () => {
-
-            // given
-            const fields = {
-                subFolder: "sub1",
-                description: "file description"
-            };
-            const files = {
-                inputFile: {
-                    size: 10,
-                    originalFilename: "original_filename.jpg",
-                    mimetype: "image/jpg"
-                }
-            };
-
-            parseStub = sinon.stub(IncomingForm.prototype, "parse").callsArgWith(1, null, fields, files);
-            const nextFake = sinon.fake();
-
-            // when
-            await formidableUploadMiddleware(requestStub, responseStub, nextFake);
-
-            // then
-            expect(requestStub.body.inputFile).not.toBeNull();
-            expect(requestStub.body.inputFile.size).toBe(files.inputFile.size);
-            expect(requestStub.body.inputFile.originalFilename).toBe(files.inputFile.originalFilename);
-            expect(requestStub.body.inputFile.mimetype).toBe(files.inputFile.mimetype);
-            expect(requestStub.body.inputFile.content).not.toBeNull();
-            expect(requestStub.body.subFolder).toBe(fields.subFolder);
-            expect(requestStub.body.description).toBe(fields.description);
-
-            sinon.assert.called(nextFake);
-        });
-
-        it("should reject parsing on error", async () => {
-
-            // given
-            parseStub = sinon.stub(IncomingForm.prototype, "parse").callsArgWith(1, new GenericError("something went wrong"), null, null);
-
-            // when
-            await formidableUploadMiddleware(requestStub, responseStub, () => {});
-
-            // then
-            sinon.assert.calledWith(responseStub.json, {message: "something went wrong"});
-            sinon.assert.calledWith(responseStub.status, HttpStatus.BAD_REQUEST);
-        });
-
-        it("should reject parsing on error (string passed as error)", async () => {
-
-            // given
-            parseStub = sinon.stub(IncomingForm.prototype, "parse").callsArgWith(1, "something went wrong", null, null);
-
-            // when
-            await formidableUploadMiddleware(requestStub, responseStub, () => {});
-
-            // then
-            sinon.assert.calledWith(responseStub.json, {message: "Invalid input file"});
-            sinon.assert.calledWith(responseStub.status, HttpStatus.BAD_REQUEST);
         });
     });
